@@ -1,85 +1,84 @@
+import 'dart:convert';
+
 import 'package:flutter/material.dart';
 import 'package:shared_preferences/shared_preferences.dart';
 
 class Settings {
-  final int workDurationMinutes;
-  final int shortBreakDurationMinutes;
-  final int longBreakDurationMinutes;
-  final int sessionsBeforeLongBreak;
+  final bool isDarkMode;
   final bool notificationsEnabled;
   final bool vibrationEnabled;
   final bool keepScreenOn;
   final String locale;
-  final bool isDarkMode;
+  final bool soundEnabled;
 
-  Settings({
-    this.workDurationMinutes = 25,
-    this.shortBreakDurationMinutes = 5,
-    this.longBreakDurationMinutes = 15,
-    this.sessionsBeforeLongBreak = 4,
-    this.notificationsEnabled = true,
-    this.vibrationEnabled = true,
-    this.keepScreenOn = true,
-    this.locale = 'auto',
-    this.isDarkMode = false,
+  const Settings({
+    required this.isDarkMode,
+    required this.notificationsEnabled,
+    required this.vibrationEnabled,
+    required this.keepScreenOn,
+    required this.locale,
+    required this.soundEnabled,
   });
 
-  factory Settings.fromPrefs(SharedPreferences prefs) {
-    return Settings(
-      workDurationMinutes: prefs.getInt('workDurationMinutes') ?? 25,
-      shortBreakDurationMinutes: prefs.getInt('shortBreakDurationMinutes') ?? 5,
-      longBreakDurationMinutes: prefs.getInt('longBreakDurationMinutes') ?? 15,
-      sessionsBeforeLongBreak: prefs.getInt('sessionsBeforeLongBreak') ?? 4,
-      notificationsEnabled: prefs.getBool('notificationsEnabled') ?? true,
-      vibrationEnabled: prefs.getBool('vibrationEnabled') ?? true,
-      keepScreenOn: prefs.getBool('keepScreenOn') ?? true,
-      locale: prefs.getString('locale') ?? 'auto',
-      isDarkMode: prefs.getBool('isDarkMode') ?? false,
-    );
-  }
-
-  Future<void> saveToPrefs(SharedPreferences prefs) async {
-    await prefs.setInt('workDurationMinutes', workDurationMinutes);
-    await prefs.setInt('shortBreakDurationMinutes', shortBreakDurationMinutes);
-    await prefs.setInt('longBreakDurationMinutes', longBreakDurationMinutes);
-    await prefs.setInt('sessionsBeforeLongBreak', sessionsBeforeLongBreak);
-    await prefs.setBool('notificationsEnabled', notificationsEnabled);
-    await prefs.setBool('vibrationEnabled', vibrationEnabled);
-    await prefs.setBool('keepScreenOn', keepScreenOn);
-    await prefs.setString('locale', locale);
-    await prefs.setBool('isDarkMode', isDarkMode);
-  }
-
   Settings copyWith({
-    int? workDurationMinutes,
-    int? shortBreakDurationMinutes,
-    int? longBreakDurationMinutes,
-    int? sessionsBeforeLongBreak,
+    bool? isDarkMode,
     bool? notificationsEnabled,
     bool? vibrationEnabled,
     bool? keepScreenOn,
     String? locale,
-    bool? isDarkMode,
+    bool? soundEnabled,
   }) {
     return Settings(
-      workDurationMinutes: workDurationMinutes ?? this.workDurationMinutes,
-      shortBreakDurationMinutes:
-          shortBreakDurationMinutes ?? this.shortBreakDurationMinutes,
-      longBreakDurationMinutes:
-          longBreakDurationMinutes ?? this.longBreakDurationMinutes,
-      sessionsBeforeLongBreak:
-          sessionsBeforeLongBreak ?? this.sessionsBeforeLongBreak,
+      isDarkMode: isDarkMode ?? this.isDarkMode,
       notificationsEnabled: notificationsEnabled ?? this.notificationsEnabled,
       vibrationEnabled: vibrationEnabled ?? this.vibrationEnabled,
       keepScreenOn: keepScreenOn ?? this.keepScreenOn,
       locale: locale ?? this.locale,
-      isDarkMode: isDarkMode ?? this.isDarkMode,
+      soundEnabled: soundEnabled ?? this.soundEnabled,
     );
+  }
+
+  Map<String, dynamic> toJson() {
+    return {
+      'isDarkMode': isDarkMode,
+      'notificationsEnabled': notificationsEnabled,
+      'vibrationEnabled': vibrationEnabled,
+      'keepScreenOn': keepScreenOn,
+      'locale': locale,
+      'soundEnabled': soundEnabled,
+    };
+  }
+
+  factory Settings.fromJson(Map<String, dynamic> json) {
+    return Settings(
+      isDarkMode: json['isDarkMode'] as bool? ?? false,
+      notificationsEnabled: json['notificationsEnabled'] as bool? ?? true,
+      vibrationEnabled: json['vibrationEnabled'] as bool? ?? true,
+      keepScreenOn: json['keepScreenOn'] as bool? ?? false,
+      locale: json['locale'] as String? ?? 'zh',
+      soundEnabled: json['soundEnabled'] as bool? ?? true,
+    );
+  }
+
+  Future<void> saveToPrefs(SharedPreferences prefs) async {
+    await prefs.setBool('isDarkMode', isDarkMode);
+    await prefs.setBool('notificationsEnabled', notificationsEnabled);
+    await prefs.setBool('vibrationEnabled', vibrationEnabled);
+    await prefs.setBool('keepScreenOn', keepScreenOn);
+    await prefs.setString('locale', locale);
+    await prefs.setBool('soundEnabled', soundEnabled);
   }
 }
 
 class SettingsModel extends ChangeNotifier {
-  Settings _settings;
+  Settings _settings = Settings(
+    isDarkMode: false,
+    notificationsEnabled: true,
+    vibrationEnabled: true,
+    keepScreenOn: false,
+    locale: 'zh',
+    soundEnabled: true,
+  );
   SharedPreferences? _prefs;
 
   SettingsModel({required Settings settings}) : _settings = settings {
@@ -109,26 +108,6 @@ class SettingsModel extends ChangeNotifier {
     notifyListeners();
   }
 
-  Future<void> setWorkDuration(int minutes) async {
-    final newSettings = _settings.copyWith(workDurationMinutes: minutes);
-    await updateSettings(newSettings);
-  }
-
-  Future<void> setShortBreakDuration(int minutes) async {
-    final newSettings = _settings.copyWith(shortBreakDurationMinutes: minutes);
-    await updateSettings(newSettings);
-  }
-
-  Future<void> setLongBreakDuration(int minutes) async {
-    final newSettings = _settings.copyWith(longBreakDurationMinutes: minutes);
-    await updateSettings(newSettings);
-  }
-
-  Future<void> setSessionsBeforeLongBreak(int sessions) async {
-    final newSettings = _settings.copyWith(sessionsBeforeLongBreak: sessions);
-    await updateSettings(newSettings);
-  }
-
   Future<void> setNotificationsEnabled(bool enabled) async {
     final newSettings = _settings.copyWith(notificationsEnabled: enabled);
     await updateSettings(newSettings);
@@ -139,8 +118,8 @@ class SettingsModel extends ChangeNotifier {
     await updateSettings(newSettings);
   }
 
-  Future<void> setKeepScreenOn(bool enabled) async {
-    final newSettings = _settings.copyWith(keepScreenOn: enabled);
+  Future<void> setSoundEnabled(bool enabled) async {
+    final newSettings = _settings.copyWith(soundEnabled: enabled);
     await updateSettings(newSettings);
   }
 
@@ -152,5 +131,30 @@ class SettingsModel extends ChangeNotifier {
   Future<void> setDarkMode(bool enabled) async {
     final newSettings = _settings.copyWith(isDarkMode: enabled);
     await updateSettings(newSettings);
+  }
+
+  Future<void> setKeepScreenOn(bool enabled) async {
+    final newSettings = _settings.copyWith(keepScreenOn: enabled);
+    await updateSettings(newSettings);
+  }
+
+  Future<void> loadSettings() async {
+    final settingsJson = _prefs?.getString('settings');
+    if (settingsJson != null) {
+      _settings = Settings.fromJson(
+        Map<String, dynamic>.from(json.decode(settingsJson)),
+      );
+    }
+    notifyListeners();
+  }
+
+  Future<void> saveSettings() async {
+    if (_prefs == null) {
+      _prefs = await SharedPreferences.getInstance();
+    }
+    await _prefs?.setString(
+      'settings',
+      json.encode(_settings.toJson()),
+    );
   }
 }
