@@ -1,4 +1,21 @@
 const Map<String, String> en = {
+  'appName': 'Tomato Clock',
+  'settings': 'Settings',
+  'notifications': 'Notifications',
+  'display': 'Display',
+  'about': 'About',
+  'enableNotifications': 'Enable Notifications',
+  'enableVibration': 'Enable Vibration',
+  'keepScreenOn': 'Keep Screen On',
+  'darkMode': 'Dark Mode',
+  'language': 'Language',
+  'version': 'Version',
+  'feedback': 'Feedback',
+  'privacyPolicy': 'Privacy Policy',
+  'termsOfService': 'Terms of Service',
+  'chinese': 'Chinese',
+  'english': 'English',
+
   'appTitle': 'Tomato Timer',
   'startButton': 'Start',
   'pauseButton': 'Pause',
@@ -20,17 +37,13 @@ const Map<String, String> en = {
   'sessionsBeforeLongBreak': 'Sessions Before Long Break',
 
   'notificationSettings': 'Notification Settings',
-  'enableNotifications': 'Enable Notifications',
-  'enableVibration': 'Enable Vibration',
+  'notificationSettingsTitle': 'Notification Settings',
 
   'displaySettings': 'Display Settings',
-  'keepScreenOn': 'Keep Screen On',
-  'darkMode': 'Dark Mode',
-  'language': 'Language',
-  'appInfo': 'App Information',
+  'displaySettingsTitle': 'Display Settings',
 
   'historyTitle': 'History',
-  'aboutTitle': 'About App',
+  'aboutTitle': 'About',
   'noHistoryMessage': 'No history records yet',
   'clearHistory': 'Clear History',
 
@@ -47,24 +60,22 @@ const Map<String, String> en = {
   'save': 'Save',
 
   'systemDefault': 'System Default',
-  'english': 'English',
-  'chinese': 'Chinese',
 
   // About page
-  'appVersion': 'App Version',
+  'appVersion': 'Version',
   'appDescription':
-      'Tomato Timer is a simple and elegant Pomodoro timer app that helps you improve your work efficiency.',
+      'A simple and elegant Pomodoro timer app that helps you stay focused and productive.',
   'developerInfo': 'Developer Information',
   'developerName': 'Wu Kun',
-  'developerEmail': 'swingcoder@gmail.com',
+  'developerEmail': 'wukun@example.com',
   'contactDeveloper': 'Contact Developer',
-  'appFeatures': 'App Features',
-  'feature1': '• Clean and elegant user interface',
-  'feature2': '• Support for work, short break, and long break timers',
-  'feature3': '• Customizable timer durations',
-  'feature4': '• Notification support',
-  'feature5': '• Dark mode support',
-  'feature6': '• English and Chinese language support',
-  'feature7': '• Vibration feedback',
-  'feature8': '• Keep screen on option',
+  'appFeatures': 'Features',
+  'feature1': '• Simple and intuitive interface',
+  'feature2': '• Customizable timer settings',
+  'feature3': '• Dark mode support',
+  'feature4': '• Sound and vibration notifications',
+  'feature5': '• Multiple language support',
+  'feature6': '• Statistics tracking',
+  'feature7': '• Screen always on option',
+  'feature8': '• Beautiful glassmorphic design',
 };
