@@ -1,4 +1,21 @@
 const Map<String, String> zh = {
+  'appName': '番茄时钟',
+  'settings': '设置',
+  'notifications': '通知',
+  'display': '显示',
+  'about': '关于',
+  'enableNotifications': '启用通知',
+  'enableVibration': '启用震动',
+  'keepScreenOn': '保持屏幕常亮',
+  'darkMode': '深色模式',
+  'language': '语言',
+  'version': '版本',
+  'feedback': '反馈',
+  'privacyPolicy': '隐私政策',
+  'termsOfService': '服务条款',
+  'chinese': '中文',
+  'english': 'English',
+
   'appTitle': '番茄计时器',
   'startButton': '开始',
   'pauseButton': '暂停',
@@ -20,13 +37,13 @@ const Map<String, String> zh = {
   'sessionsBeforeLongBreak': '长休息前工作阶段数',
 
   'notificationSettings': '通知设置',
-  'enableNotifications': '启用通知',
-  'enableVibration': '启用振动',
+  'notificationSettingsTitle': '通知设置',
+  'notificationSettingsDescription': '设置通知选项',
 
   'displaySettings': '显示设置',
-  'keepScreenOn': '保持屏幕常亮',
-  'darkMode': '暗黑模式',
-  'language': '语言',
+  'displaySettingsTitle': '显示设置',
+  'displaySettingsDescription': '设置显示选项',
+
   'appInfo': '应用信息',
 
   'historyTitle': '历史记录',
@@ -47,23 +64,21 @@ const Map<String, String> zh = {
   'save': '保存',
 
   'systemDefault': '系统默认',
-  'english': '英文',
-  'chinese': '中文',
 
   // 关于页面
   'appVersion': '应用版本',
-  'appDescription': '番茄计时器是一个简单而优雅的番茄工作法应用，帮助你提高工作效率。',
+  'appDescription': '一个简单优雅的番茄钟应用，帮助你保持专注和提高效率。',
   'developerInfo': '开发者信息',
   'developerName': 'swingcoder',
   'developerEmail': 'swingcoder@gmail.com',
   'contactDeveloper': '联系开发者',
-  'appFeatures': '应用特点',
-  'feature1': '• 简洁优雅的用户界面',
-  'feature2': '• 支持工作、短休息和长休息计时',
-  'feature3': '• 自定义计时时长',
-  'feature4': '• 支持通知提醒',
-  'feature5': '• 支持深色模式',
-  'feature6': '• 支持中英文切换',
-  'feature7': '• 支持振动反馈',
-  'feature8': '• 支持保持屏幕常亮',
+  'appFeatures': '功能特点',
+  'feature1': '• 简单直观的界面',
+  'feature2': '• 可自定义的计时器设置',
+  'feature3': '• 深色模式支持',
+  'feature4': '• 声音和震动通知',
+  'feature5': '• 多语言支持',
+  'feature6': '• 统计跟踪',
+  'feature7': '• 屏幕常亮选项',
+  'feature8': '• 精美的毛玻璃设计',
 };
